+++ packages/app-builder/src/components/Scenario/AstBuilder/AstBuilderNode/Operator.tsx
@@ -1,8 +1,6 @@
-import {
-  getOperatorName,
-  type OperatorFunction,
-  undefinedAstNodeName,
-} from '@app-builder/models/editable-operators';
+import { undefinedAstNodeName } from '@app-builder/models/astNode/ast-node';
+import { getOperatorName } from '@app-builder/models/getOperatorName';
+import { type OperatorFunction } from '@app-builder/models/modale-operators';
 import { Trigger, Value } from '@radix-ui/react-select';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { forwardRef } from 'react';
